+++ src/main/java/com/translator/TranslatorPlugin.java
@@ -35,6 +35,7 @@ import net.runelite.api.events.MenuOpened;
 import net.runelite.api.widgets.ComponentID;
 import net.runelite.api.widgets.Widget;
 import net.runelite.api.ItemComposition;
+import net.runelite.client.RuneLite;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.eventbus.Subscribe;
 import net.runelite.client.events.ConfigChanged;
@@ -47,6 +48,8 @@ import org.slf4j.LoggerFactory;
 import javax.inject.Inject;
 import java.io.*;
 import java.net.*;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -61,7 +64,6 @@ import java.util.concurrent.TimeUnit;
 )
 public class TranslatorPlugin extends Plugin
 {
-
 	private static final Logger log = LoggerFactory.getLogger(TranslatorPlugin.class);
 	@Inject
     private TranslatorConfig config;
@@ -75,18 +77,47 @@ public class TranslatorPlugin extends Plugin
 	private TranslatorAPI api;
 	private ScheduledExecutorService executor;
 
-    private HashMap<String, String> itemsMap;
-    private HashMap<String, String> npcMap;
-    private HashMap<String, String> objectMap;
-    private HashMap<String, String> dialogueMap;
+    private HashMap<String, String> itemsMap = new HashMap<>();
+    private HashMap<String, String> npcMap = new HashMap<>();
+    private HashMap<String, String> objectMap = new HashMap<>();
+    private HashMap<String, String> dialogueMap = new HashMap<>();
     private final HashSet<String> dialoguesToSend = new HashSet<String>();
     private Widget prevTickWidget;
     private Actor actor;
 
+	private void loadTranslation(
+		TranslatorAPI.TranslationFileType type,
+		TranslatorConfig.Language lang
+	) {
+		executor.execute(() -> {
+			try {
+				switch (type) {
+					case dialogue:
+						dialogueMap = api.getTranslationMap(type, lang);
+						break;
+					case npc:
+						npcMap = api.getTranslationMap(type, lang);
+						break;
+					case object:
+						objectMap = api.getTranslationMap(type, lang);
+						break;
+					case items:
+						itemsMap = api.getTranslationMap(type, lang);
+						break;
+				}
+			} catch (Exception e) {
+				log.error("failed to load translation '{}:{}'", type, lang, e);
+			}
+		});
+	}
 
     @Override
     protected void startUp() throws Exception {
 		executor = Executors.newSingleThreadScheduledExecutor();
+		loadTranslation(TranslatorAPI.TranslationFileType.dialogue, config.selectLanguage());
+		loadTranslation(TranslatorAPI.TranslationFileType.npc, config.selectLanguage());
+		loadTranslation(TranslatorAPI.TranslationFileType.items, config.selectLanguage());
+		loadTranslation(TranslatorAPI.TranslationFileType.object, config.selectLanguage());
 
 		if (config.enableTextCapture()) {
 			executor.scheduleAtFixedRate(
@@ -96,8 +127,6 @@ public class TranslatorPlugin extends Plugin
 				TimeUnit.MINUTES
 			);
 		}
-
-        updateLanguage();
     }
 
 	@Override
@@ -107,8 +136,15 @@ public class TranslatorPlugin extends Plugin
 
 	@Subscribe
 	public void onConfigChanged(ConfigChanged event) {
-		if (event.getGroup().equals("translator")) {
-			updateLanguage();
+		if (!event.getGroup().equals("translator")) {
+			return;
+		}
+
+		if (event.getKey().equals("SelectLanguage")) {
+			loadTranslation(TranslatorAPI.TranslationFileType.dialogue, config.selectLanguage());
+			loadTranslation(TranslatorAPI.TranslationFileType.npc, config.selectLanguage());
+			loadTranslation(TranslatorAPI.TranslationFileType.items, config.selectLanguage());
+			loadTranslation(TranslatorAPI.TranslationFileType.object, config.selectLanguage());
 		}
 
 		if (event.getKey().equals("enableOSRSTextCollection")) {
@@ -124,50 +160,6 @@ public class TranslatorPlugin extends Plugin
 			}
 		}
 	}
-
-    public HashMap<String, String> parseDialogue(String filepath) {
-        HashMap<String, String> words = new HashMap<String, String>();
-
-        try (BufferedReader br = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream(filepath)))) {
-            String line;
-            while ((line = br.readLine()) != null) {
-                String[] temp = line.split(";");
-                if (temp.length > 1) {
-                    words.put(temp[0], temp[1]);
-                }
-            }
-            return words;
-        } catch (IOException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    public HashMap<String, String> parse(String filepath) {
-        HashMap<String, String> words = new HashMap<String, String>();
-
-        try (BufferedReader br = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream(filepath)))) {
-            String line;
-            while ((line = br.readLine()) != null) {
-                String[] temp = line.split(",");
-                if (temp.length > 2) {
-                    words.put(temp[0], temp[2]);
-                }
-            }
-            return words;
-        } catch (IOException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    private void updateLanguage()
-    {
-        itemsMap = parse("/"+ config.selectLanguage() +"_items.txt");
-        npcMap = parse("/"+ config.selectLanguage() +"_npc.txt");
-        objectMap = parse("/"+ config.selectLanguage() +"_object.txt");
-        dialogueMap = parseDialogue("/"+ config.selectLanguage() +"_dialogue.txt");
-    }
 
     @Subscribe
     public void onMenuOpened(MenuOpened event)
@@ -314,8 +306,11 @@ public class TranslatorPlugin extends Plugin
 		}
 	}
 
-    private void checkWidgetDialogs()
-    {
+    private void checkWidgetDialogs() {
+		if (!config.translateWidgets()) {
+			return;
+		}
+
         Widget widget;
         if (client.getWidget(ComponentID.DIALOG_NPC_TEXT) != null) {
             widget = client.getWidget(ComponentID.DIALOG_NPC_TEXT);