+++ CHANGELOG.md
@@ -5,6 +5,41 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Dialogue translation can now be disabled.
+- Translation files are now loaded dynamically from (iolave/osrs-translations)[github.com/iolave/osrs-translations].
+
+### Changed
+- Renamed `SelectLanguage` enum to `Language`.
+
+## [v1.2.0] - 2026-01-14
+
+### Added
+- Added Portuguese support.
+- Portuguese translation for dialogues, items, NPCs, and objects.
+
+### Changed
+- Reduced Finnish translation file sizes.
+
+### Fixed
+- Incorrect schedule time on plugin change.
+
+### Removed
+- Removed unused resource files.
+
+## [v1.1.0] - 2026-01-07
+
+### Added
+- Added an option to enable/disable dialogue options translation.
+- Added opt-in data collection via runelite-translator-api.
+- Added spanish support.
+- Data collection is now opt-in.
+
+### Removed
+- Removed sys outs.
+
 ## [v1.0.0] - 2025-12-03
 
 ### Added